backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
httpx==0.24.1
firebase-admin==6.2.0
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6
pandas==2.1.1
numpy==1.26.0
websockets==11.0.3
scikit-learn==1.3.1